Fix warning: Remove superflous casts in getScrollDelta
../src/wcmCommon.c: In function 'getScrollDelta':
../src/wcmCommon.c:333:13: warning: cast from function call of type 'double' to non-matching type 'int' [-Wbad-function-cast]
../src/wcmCommon.c:334:9: warning: cast from function call of type 'double' to non-matching type 'int' [-Wbad-function-cast]
../src/wcmCommon.c:335:10: warning: cast from function call of type 'double' to non-matching type 'int' [-Wbad-function-cast]

+++ src/wcmCommon.c
@@ -330,9 +330,9 @@ static int getScrollDelta(int current, int old, int wrap, int flags)
 
 	if (flags & AXIS_BITWISE)
 	{
-		current = (int)log2((current << 1) | 0x01);
-		old = (int)log2((old << 1) | 0x01);
-		wrap = (int)log2((wrap << 1) | 0x01);
+		current = log2((current << 1) | 0x01);
+		old = log2((old << 1) | 0x01);
+		wrap = log2((wrap << 1) | 0x01);
 	}
 
 	delta = current - old;
